+++ CHANGELOG.md
@@ -39,6 +39,7 @@
   - Fix add layer with the correct beforeId on update `PaintProperty` workaround for [maplibre issue 3001](https://github.com/maplibre/maplibre-gl-js/issues/3001).
   - Fix workaround that tried to get beforeId after layer was removed. Now use workaround only if `map.getTerrain()` is not not null.
   - Fix set `layout.visibility` in style of `VectorLayer` based on `VectorLayer.visible`.
+  - Fix set `VectorLayer.attribution` from `VectorLayer.options.style` for TMS.
 
 * **@dlr-eoc/map-cesium:**
   - Fix get entity.name for infoBox in none GeoJson layers